fmcsa_system/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Scheduling
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# API Features
python-multipart==0.0.6
aiofiles==23.2.1
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1